requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
pathlib>=1.0.1
tqdm>=4.65.0
python-magic>=0.4.27
pandas>=2.0.0 